refactor: Move config stuff to gallia.config
