+++ src/gallia/cli.py
@@ -7,7 +7,6 @@
 import argparse
 import logging
 import os
-import subprocess
 import sys
 from importlib.metadata import entry_points, version
 from pathlib import Path
@@ -15,9 +14,8 @@ from pprint import pprint
 from typing import Any, Optional
 
 import argcomplete
-import tomlkit
-from xdg import xdg_config_dirs
 
+from gallia.config import load_config_file
 from gallia.command import BaseCommand
 from gallia.commands.discover.uds.doip import DoIPDiscoverer
 from gallia.commands.discover.uds.isotp import IsotpDiscoverer
@@ -75,49 +73,6 @@ registry: list[type[BaseCommand]] = [
 # This can be annotated once recursive types are supported by mypy.
 # https://github.com/python/mypy/issues/731
 PARSERS: dict[str, Any] = {}
-
-
-def get_git_root() -> Optional[Path]:
-    try:
-        p = subprocess.run(
-            ["git", "rev-parse", "--show-toplevel"],
-            capture_output=True,
-            check=True,
-        )
-    except subprocess.SubprocessError:
-        return None
-
-    return Path(p.stdout.decode().strip())
-
-
-def get_config_dirs() -> list[Path]:
-    dirs = xdg_config_dirs()
-    git_root = get_git_root()
-    cwd = Path.cwd()
-    if git_root is not None:
-        return [cwd, git_root] + dirs
-    return [cwd] + dirs
-
-
-def search_config() -> Optional[Path]:
-    if (s := os.getenv("GALLIA_CONFIG")) is not None:
-        if (path := Path(s)).exists():
-            return path
-        else:
-            raise FileNotFoundError(s)
-
-    for dir_ in get_config_dirs():
-        if (path := dir_.joinpath("gallia.toml")).exists():
-            return path
-
-    return None
-
-
-def load_config_file() -> tuple[dict[str, Any], Optional[Path]]:
-    if (path := search_config()) is not None:
-        raw_toml = path.read_text()
-        return tomlkit.loads(raw_toml), path
-    return {}, None
 
 
 def load_cli_commands() -> None:
